rest-app/requirements.txt
uvicorn==0.15.0
fastapi
cassandra-driver